+++ src/components/Modal/MapContainer.tsx
@@ -1,41 +1,13 @@
-import React, { useRef, useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import { Map, ZoomControl, MapMarker } from 'react-kakao-maps-sdk';
-import { addressAction } from '../../redux/features/addressSlice';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../../redux/store';
-import { toggleModalShow } from '../../redux/features/toggleSlice';
 import { useCalendar } from '../../hooks/modal/useCalendar';
+import { useMap } from '../../hooks/map/useMap';
 import Modal from './Modal';
 
 const MapContainer = () => {
-  const mapRef = useRef(null);
-  const dispatch = useDispatch();
-  const nowPosition = useSelector((state: RootState) => state.persistReducered.position);
-  const [position, setPosition] = useState({ lat: 0, lng: 0 });
-  const [address, setAddress] = useState<string>();
   const { handleToggleModal } = useCalendar('');
-
-  useEffect(() => {
-    getAddress(position.lat, position.lng);
-  }, [position]);
-
-  const getAddress = (lat: number, lng: number) => {
-    const geocoder = new window.kakao.maps.services.Geocoder();
-    const coord = new window.kakao.maps.LatLng(lat, lng);
-    const callback = function (result: any, status: string) {
-      if (status === window.kakao.maps.services.Status.OK) {
-        const arr = { ...result };
-        const arr1 = arr[0].address.address_name;
-        setAddress(arr1);
-      }
-    };
-    geocoder.coord2Address(coord.getLng(), coord.getLat(), callback);
-  };
-
-  const handleSendAddress = () => {
-    dispatch(addressAction({ address: address, lat: position.lat, lng: position.lng }));
-    dispatch(toggleModalShow());
-  };
+  const { mapRef, nowPosition, position, setPosition, address, handleSendAddress, loadAddress } = useMap('');
+  loadAddress();
 
   return (
     <Modal onClickToggleModal={handleToggleModal}>
@@ -52,27 +24,24 @@ const MapContainer = () => {
         }
       >
         <ZoomControl position={window.kakao.maps.ControlPosition.TOPRRIGHT} />
-
-        <>
-          <MapMarker position={{ lat: position.lat, lng: position.lng }} />
-          <div className='w-full h-[50px] flex justify-center items-center bg-[#FCFCFC] mt-3'>{address}</div>
-          <span className='flex flex-row items-center gap-5 mt-3'>
-            <button
-              className='w-[82px] h-[45px] bg-[#FFF] text-[16px] text-[#38393C] leading-[19px] border border-[#9A9B9F]
+        <MapMarker position={{ lat: position.lat, lng: position.lng }} />
+        <div className='w-full h-[50px] flex justify-center items-center bg-[#FCFCFC] mt-3'>{address}</div>
+        <span className='flex flex-row items-center gap-5 mt-3'>
+          <button
+            className='w-[82px] h-[45px] bg-[#FFF] text-[16px] text-[#38393C] leading-[19px] border border-[#9A9B9F]
             font-medium text-center rounded-[8px]'
-              onClick={handleToggleModal}
-            >
-              닫기
-            </button>
-            <button
-              className='w-[181px] h-[45px] bg-matchgi-btnblue text-[16px] text-[#FFF] leading-[19px]
+            onClick={handleToggleModal}
+          >
+            닫기
+          </button>
+          <button
+            className='w-[181px] h-[45px] bg-matchgi-btnblue text-[16px] text-[#FFF] leading-[19px]
             font-medium text-center rounded-[8px]'
-              onClick={handleSendAddress}
-            >
-              선택
-            </button>
-          </span>
-        </>
+            onClick={handleSendAddress}
+          >
+            선택
+          </button>
+        </span>
       </Map>
     </Modal>
   );
