+++ src/util/constant/makerTable.ts
@@ -1,0 +1,11 @@
+import { ImageType } from '../../typings';
+
+export const MarkerObj: ImageType = {
+  축구: '/assets/images/map/soccer.svg',
+  농구: '/assets/images/map/basketball.svg',
+  기타: '/assets/images/map/etc.svg',
+  볼링: '/assets/images/map/bowling.svg',
+  배드민턴: '/assets/images/map/badminton.svg',
+  테니스: '/assets/images/map/tennis.svg',
+  당구: '/assets/images/map/bil.svg',
+};
